+++ Assets/Player.meta
@@ -4,7 +4,6 @@ folderAsset: yes
 timeCreated: 1516012667
 licenseType: Free
 DefaultImporter:
-  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
