+++ Assets/Prefab/Player.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 9404f57a6a8809e4b95dc451c35a641b
+timeCreated: 1516872262
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
