+++ Assets/good knuckles.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 1fb368bae657de249ae3e678dbdedc60
+guid: f77a0e87c6a25934fbae7f5220e022b0
 folderAsset: yes
-timeCreated: 1516870628
+timeCreated: 1517215109
 licenseType: Free
 DefaultImporter:
   userData:
