+++ Assets/good knuckles/Materials.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 24302f54a76ad8441b50ba49cc0d1f71
+guid: 911ff8e66b9ee87489e7470bc254264b
 folderAsset: yes
-timeCreated: 1516015308
+timeCreated: 1517215367
 licenseType: Free
 DefaultImporter:
   userData:
