+++ Assets/good knuckles/Materials/No Name.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 669555df0e6e63340b6e017c0f35754a
+timeCreated: 1517215367
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
