+++ Assets/good knuckles/Sphere_1.controller.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: fae145273250bb642b548fc979ca6fd4
+timeCreated: 1517216104
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
