+++ Assets/good knuckles/goodknuckles textures.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 4072817239df60c4cae0d1890ad9ba08
+timeCreated: 1517215116
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
