+++ Assets/good knuckles/head knuckles good.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: a6780c07b02f3894fbfeb6011eabf248
+timeCreated: 1517215116
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
