+++ Assets/good knuckles/run knuckles good test.controller
@@ -1,0 +1,11 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!91 &9100000
+AnimatorController:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_Name: run knuckles good test
+  serializedVersion: 5
+  m_AnimatorParameters: []
+  m_AnimatorLayers: []
