+++ Assets/good knuckles/run knuckles good test.controller.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 2d5823d692869da4298c0eb8c654f03f
+timeCreated: 1517216145
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
