+++ Assets/good knuckles/run knuckles good.prefab.meta
@@ -1,9 +1,9 @@
 fileFormatVersion: 2
-guid: e1e84bea59e04c7418abd40daad74162
-timeCreated: 1516870628
+guid: 85022c285d447b94c8914402f2d915d6
+timeCreated: 1517215116
 licenseType: Free
 NativeFormatImporter:
-  mainObjectFileID: 2100000
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
